Revamp the docs in `expression::helper_types`
I've converted everything to links. The function links are `../dsl/fn`
instead of just `fn`, because these types are rendered both in the `dsl`
module and the `helper_types` module.

I've added a bunch of missing helper types, restructured the module so
that everything is ordered the same as they're defined on the traits
involved, to make it easier for me to ensure we have a helper type for
every method in `expression_methods`.

I've also added a bunch of helper types for the various bare functions,
and converted `not` to use our convention for function helper types.